+++ pakyow-support/CHANGELOG.md
@@ -1,5 +1,8 @@
 # v1.1.0 (unreleased)
 
+  * `chg` **Improve isolated module inheritance.**
+    - [Pull Request #366][pr-366]
+
   * `chg` **Register pipeline actions during initialization.**
     - [Pull Request #365][pr-365]
 
@@ -83,6 +86,7 @@
     *Related links:*
     - [Pull Request #364][pr-364]
 
+[pr-366]: https://github.com/pakyow/pakyow/pull/366
 [pr-365]: https://github.com/pakyow/pakyow/pull/365
 [pr-364]: https://github.com/pakyow/pakyow/pull/364
 [pr-363]: https://github.com/pakyow/pakyow/pull/363